refactor: extract shared OAuth helpers and deduplicate utilities
